// src/types/vaildate/InputGuideType.ts
export interface InputGuideType {
  name: string;
  message: string;
}

export interface CategoryType {
  input: string;
}

export interface NameType {
  input: string;
  maxLength: number;
}

export interface DistanceType {
  input: string;
}

export interface DescriptionType {
  input: string;
  maxLength: number;
}

export interface LinkType {
  input: string;
}